backend01/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.4

# Bioinformatics
PyVCF3==1.0.4
pandas==2.2.3

# Database
aiosqlite==0.20.0

# AI / LLM
google-generativeai==0.8.4
langchain==0.3.14
langchain-google-genai==2.0.8
llama-index==0.12.3
pinecone-client==5.0.1
